+++ fili-core/src/test/groovy/com/yahoo/bard/webservice/data/metric/TemplateDruidQuerySpec.groovy
@@ -22,9 +22,9 @@ class TemplateDruidQuerySpec extends Specification {
             Set<Aggregation> aggs = []
             Set<PostAggregation> postAggs = []
 
-            TemplateDruidQuery q1 = new TemplateDruidQuery(aggs, postAggs, (ZonelessTimeGrain) null)
-            TemplateDruidQuery q2 = new TemplateDruidQuery(aggs, postAggs, q1, (ZonelessTimeGrain) null)
-            TemplateDruidQuery q3 = new TemplateDruidQuery(aggs, postAggs, q2, (ZonelessTimeGrain) null)
+            TemplateDruidQuery q1 = new TemplateDruidQuery(aggs, postAggs)
+            TemplateDruidQuery q2 = new TemplateDruidQuery(aggs, postAggs, q1)
+            TemplateDruidQuery q3 = new TemplateDruidQuery(aggs, postAggs, q2)
 
         expect:
             q1.depth() == 1
@@ -37,17 +37,17 @@ class TemplateDruidQuerySpec extends Specification {
             Set<Aggregation> aggs = []
             Set<PostAggregation> postAggs = []
 
-            TemplateDruidQuery q1 = new TemplateDruidQuery(aggs, postAggs, (ZonelessTimeGrain) null)
-            TemplateDruidQuery q2 = new TemplateDruidQuery(aggs, postAggs, q1, (ZonelessTimeGrain) null)
-            TemplateDruidQuery q3 = new TemplateDruidQuery(aggs, postAggs, q2, (ZonelessTimeGrain) null)
+            TemplateDruidQuery q1 = new TemplateDruidQuery(aggs, postAggs)
+            TemplateDruidQuery q2 = new TemplateDruidQuery(aggs, postAggs, q1)
+            TemplateDruidQuery q3 = new TemplateDruidQuery(aggs, postAggs, q2)
 
         expect:
             q1.isNested() == false
             q2.isNested() == true
             q3.isNested() == true
-            q2.getInnerQuery().isNested() == false
-            q3.getInnerQuery().isNested() == true
-            q3.getInnerQuery().getInnerQuery().isNested() == false
+            q2.getInnerQuery().get().isNested() == false
+            q3.getInnerQuery().get().isNested() == true
+            q3.getInnerQuery().get().getInnerQuery().get().isNested() == false
             q1.isTimeGrainValid() == true
     }
 
@@ -60,7 +60,7 @@ class TemplateDruidQuerySpec extends Specification {
             PostAggregation postagg3 = new ArithmeticPostAggregation("field3", ArithmeticPostAggregationFunction.PLUS,
                     [postagg1, postagg2]
             )
-            TemplateDruidQuery q1 = new TemplateDruidQuery([agg1,agg2] as Set, [postagg3] as Set, (ZonelessTimeGrain) null)
+            TemplateDruidQuery q1 = new TemplateDruidQuery([agg1,agg2] as Set, [postagg3] as Set)
             TemplateDruidQuery q2 = q1.nest()
 
         expect:
@@ -68,8 +68,8 @@ class TemplateDruidQuerySpec extends Specification {
             q2.getAggregations().sort() == [agg1, agg2].sort()
             q2.getPostAggregations().sort() == [postagg3].sort()
             q2.getInnerQuery() != null
-            q2.getInnerQuery().getAggregations().sort() == [agg1, agg2].sort()
-            q2.getInnerQuery().getPostAggregations().isEmpty()
+            q2.getInnerQuery().get().getAggregations().sort() == [agg1, agg2].sort()
+            q2.getInnerQuery().get().getPostAggregations().isEmpty()
     }
 
     def "verify q1.merge(q2) equals merged"() {
@@ -89,10 +89,10 @@ class TemplateDruidQuerySpec extends Specification {
             PostAggregation nested_postagg1 = new ConstantPostAggregation("field6", 100)
             PostAggregation q2_postagg1 = new ConstantPostAggregation("field7", 100)
 
-            TemplateDruidQuery nested = new TemplateDruidQuery([nested_agg1] as Set, [nested_postagg1] as Set, (ZonelessTimeGrain) null)
-            TemplateDruidQuery q1 = new TemplateDruidQuery([q1_agg1,q1_agg2] as Set, [q1_postagg1] as Set, nested, (ZonelessTimeGrain) null)
+            TemplateDruidQuery nested = new TemplateDruidQuery([nested_agg1] as Set, [nested_postagg1] as Set)
+            TemplateDruidQuery q1 = new TemplateDruidQuery([q1_agg1,q1_agg2] as Set, [q1_postagg1] as Set, nested)
 
-            TemplateDruidQuery q2 = new TemplateDruidQuery([q2_agg1] as Set, [q2_postagg1] as Set, (ZonelessTimeGrain) null)
+            TemplateDruidQuery q2 = new TemplateDruidQuery([q2_agg1] as Set, [q2_postagg1] as Set)
 
             TemplateDruidQuery merged = q1.merge(q2)
 
@@ -101,8 +101,8 @@ class TemplateDruidQuerySpec extends Specification {
             merged.getAggregations().sort() == [q1_agg1, q1_agg2, new LongSumAggregation("foo", "foo")].sort()
             merged.getPostAggregations().sort() == [q1_postagg1, q2_postagg1].sort()
             merged.depth() == 2
-            merged.getInnerQuery().getAggregations().sort() == [q2_agg1, nested_agg1].sort()
-            merged.getInnerQuery().getPostAggregations().sort() == [nested_postagg1]
+            merged.getInnerQuery().get().getAggregations().sort() == [q2_agg1, nested_agg1].sort()
+            merged.getInnerQuery().get().getPostAggregations().sort() == [nested_postagg1]
 
             q1.merge(q2) == q2.merge(q1)
     }
@@ -115,9 +115,9 @@ class TemplateDruidQuerySpec extends Specification {
 
             Aggregation q2_agg1 = new LongSumAggregation("duplicate", "duplicate")
 
-            TemplateDruidQuery nested = new TemplateDruidQuery([nested_agg1] as Set, [] as Set, (ZonelessTimeGrain) null)
-            TemplateDruidQuery q1 = new TemplateDruidQuery([q1_agg1, q1_agg2] as Set, [] as Set, nested, (ZonelessTimeGrain) null)
-            TemplateDruidQuery q2 = new TemplateDruidQuery([q2_agg1] as Set, [] as Set, (ZonelessTimeGrain) null)
+            TemplateDruidQuery nested = new TemplateDruidQuery([nested_agg1] as Set, [] as Set)
+            TemplateDruidQuery q1 = new TemplateDruidQuery([q1_agg1, q1_agg2] as Set, [] as Set, nested)
+            TemplateDruidQuery q2 = new TemplateDruidQuery([q2_agg1] as Set, [] as Set)
 
         when:
             q1.merge(q2)
@@ -144,9 +144,9 @@ class TemplateDruidQuerySpec extends Specification {
             PostAggregation nested_postagg1 = new ConstantPostAggregation("field5", 100)
             PostAggregation q2_postagg1 = new ConstantPostAggregation("duplicate", 100)
 
-            TemplateDruidQuery nested = new TemplateDruidQuery([nested_agg1] as Set, [nested_postagg1] as Set, (ZonelessTimeGrain) null)
-            TemplateDruidQuery q1 = new TemplateDruidQuery([q1_agg1, q1_agg2] as Set, [q1_postagg1] as Set, nested, (ZonelessTimeGrain) null)
-            TemplateDruidQuery q2 = new TemplateDruidQuery([q2_agg1] as Set, [q2_postagg1] as Set, (ZonelessTimeGrain) null)
+            TemplateDruidQuery nested = new TemplateDruidQuery([nested_agg1] as Set, [nested_postagg1] as Set)
+            TemplateDruidQuery q1 = new TemplateDruidQuery([q1_agg1, q1_agg2] as Set, [q1_postagg1] as Set, nested)
+            TemplateDruidQuery q2 = new TemplateDruidQuery([q2_agg1] as Set, [q2_postagg1] as Set)
 
         when:
             q1.merge(q2)
@@ -166,7 +166,7 @@ class TemplateDruidQuerySpec extends Specification {
             PostAggregation p3 = new ConstantPostAggregation("dup", 100)
 
         when:
-            new TemplateDruidQuery([a1,a2,a3] as Set, [p1,p2,p3] as Set, (ZonelessTimeGrain) null)
+            new TemplateDruidQuery([a1,a2,a3] as Set, [p1,p2,p3] as Set)
 
         then:
             thrown(IllegalArgumentException)