+++ fili-wikipedia-example/src/test/groovy/com/yahoo/wiki/webservice/web/endpoints/TablesServletSpec.groovy
@@ -38,21 +38,21 @@ class TablesServletSpec extends Specification {
                                                     "granularity":"hour",
                                                     "name":"$tableName",
                                                     "longName":"$tableName",
-                                                    "uri":"http://localhost:9998/tables/$tableName/hour"
+                                                    "uri":"http://localhost:${jerseyTestBinder.getHarness().getPort()}/tables/$tableName/hour"
                                             },
                                             {
                                                     "category":"General",
                                                     "granularity":"day",
                                                     "name":"$tableName",
                                                     "longName":"$tableName",
-                                                    "uri":"http://localhost:9998/tables/$tableName/day"
+                                                    "uri":"http://localhost:${jerseyTestBinder.getHarness().getPort()}/tables/$tableName/day"
                                             },
                                             {
                                                     "category":"General",
                                                     "granularity":"all",
                                                     "name":"$tableName",
                                                     "longName":"$tableName",
-                                                    "uri":"http://localhost:9998/tables/$tableName/all"
+                                                    "uri":"http://localhost:${jerseyTestBinder.getHarness().getPort()}/tables/$tableName/all"
                                             }
                                         ]
                                     }"""
@@ -71,21 +71,21 @@ class TablesServletSpec extends Specification {
                                                 "granularity":"hour",
                                                 "name":"$tableName",
                                                 "longName":"$tableName",
-                                                "uri":"http://localhost:9998/tables/$tableName/hour"
+                                                "uri":"http://localhost:${jerseyTestBinder.getHarness().getPort()}/tables/$tableName/hour"
                                             },
                                             {
                                                 "category":"General",
                                                 "granularity":"day",
                                                 "name":"$tableName",
                                                 "longName":"$tableName",
-                                                "uri":"http://localhost:9998/tables/$tableName/day"
+                                                "uri":"http://localhost:${jerseyTestBinder.getHarness().getPort()}/tables/$tableName/day"
                                             },
                                             {
                                                 "category":"General",
                                                 "granularity":"all",
                                                 "name":"$tableName",
                                                 "longName":"$tableName",
-                                                "uri":"http://localhost:9998/tables/$tableName/all"
+                                                "uri":"http://localhost:${jerseyTestBinder.getHarness().getPort()}/tables/$tableName/all"
                                             }
                                         ]
                                     }"""
@@ -121,7 +121,7 @@ class TablesServletSpec extends Specification {
                                                     "name": "$it",
                                                     "longName": "wiki $it",
                                                     "cardinality": 0,
-                                                    "uri": "http://localhost:9998/dimensions/$it"
+                                                    "uri": "http://localhost:${jerseyTestBinder.getHarness().getPort()}/dimensions/$it"
                                                     }"""
                                                 }
                                                 .join(',')
@@ -134,7 +134,7 @@ class TablesServletSpec extends Specification {
                                                         "category": "General",
                                                         "name": "$it",
                                                         "longName": "$it",
-                                                        "uri": "http://localhost:9998/metrics/$it"
+                                                        "uri": "http://localhost:${jerseyTestBinder.getHarness().getPort()}/metrics/$it"
                                                     }"""
                                                 }
                                                 .join(',')